+++ tf/parameters.py
@@ -8,7 +8,7 @@ import sys
 from zipfile import ZIP_DEFLATED
 
 
-VERSION = '9.4.4'
+VERSION = '9.5.0'
 """Program version.
 
 This value is under control of the update process, as run by